git-svn-id: https://dng.biozentrum.unibas.ch/svn/openstructure/branches/new_shell@2368 5a81b35b-ba03-0410-adc8-b2c5c5119f08

+++ modules/io/src/mol/entity_io_mae_handler.cc
@@ -133,7 +133,7 @@ void MAEReader::Import(mol::EntityHandle& ent)
           } else {
             // parsing atom line
             std::vector<std::string> tokens=tokenize(line);
-            for(int i=0;i<tokens.size();++i) {
+            for(size_t i=0;i<tokens.size();++i) {
               LOG_DUMP( "[" << tokens[i] << "] ");
             }
             LOGN_DUMP("");
